week09/asr/requirements.txt
pytorch-lightning==1.5.9
hydra-core=1.1.1
editdistance
torchaudio
torch
pandas
sentencepiece==0.1.96
